added more info about weather, new ui components and pictures

+++ app/src/main/java/com/mentarey/easyweather/data/weather/model/Sys.kt
@@ -10,9 +10,9 @@ data class Sys(
     @SerializedName("message")
     val message: Double,
     @SerializedName("sunrise")
-    val sunrise: Int,
+    val sunrise: Long,
     @SerializedName("sunset")
-    val sunset: Int,
+    val sunset: Long,
     @SerializedName("type")
     val type: Int
 )
